Tests structure refactoring
